utils/ reorganize dependencies and clean up tests
