+++ .changeset/eleven-bugs-greet.md
@@ -1,0 +1,5 @@
+---
+"@mud-classic/utils": patch
+---
+
+reorganize packages and clean up testing
