+++ package.json
@@ -14,19 +14,31 @@
     "nuke": "pnpm -r nuke && rimraf pnpm-lock.yaml node_modules",
     "test": "pnpm -r test"
   },
+  "dependencies": {
+    "ethers": "^5.5.0",
+    "mobx": "^6.5.0",
+    "rxjs": "^7.5.5"
+  },
   "devDependencies": {
+    "@babel/core": "^7.23.9",
+    "@babel/preset-env": "^7.23.9",
     "@changesets/cli": "^2.27.1",
+    "@types/jest": "^29.5.11",
     "@typescript-eslint/eslint-plugin": "^5.23.0",
     "@typescript-eslint/parser": "^5.23.0",
+    "babel-jest": "^27.0.0",
     "eslint": "^8.15.0",
+    "jest": "^29.7.0",
     "jq-cli-wrapper": "^1.6.1",
     "lint-staged": ">=10",
     "prettier": "^2.6.2",
     "retypeapp": "^2.4.0",
     "rimraf": "^3.0.2",
     "run-pty": "^3.0.0",
+    "ts-jest": "^29.1.2",
+    "tslib": "^2.3.1",
     "tsup": "^8.0.1",
-    "typescript": ">=3.0.0"
+    "typescript": "^4.5.0"
   },
   "lint-staged": {
     "*.ts": "eslint --cache --fix",
