+++ packages/utils/jest.config.js
@@ -1,5 +1,9 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-export default  {
+/** @type {import('ts-jest').Options} */
+export default {
   preset: "ts-jest",
+  roots: ["tests"],
   testEnvironment: "node",
+  transform: {
+    '^.+\\.(js|jsx)$': 'babel-jest',
+  },
 };
