+++ packages/utils/package.json
@@ -15,26 +15,16 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "prepare": "pnpm build",
-    "build": "tsup src/index.ts --format cjs,esm --dts-resolve --minify --clean --sourcemap --legacy-output",
+    "build": "tsup ./src",
     "dev": "tsup src/index.ts --format cjs,esm --dts-resolve --watch --sourcemap --legacy-output",
     "lint": "eslint . --ext .ts",
     "nuke": "rimraf node_modules/ dist/",
     "release": "pnpm run build && changeset publish",
     "test": "pnpm run build && jest --coverage"
   },
-  "dependencies": {
-    "ethers": "^5.5.0",
-    "mobx": "^6.5.0",
-    "rxjs": "^7.5.5"
-  },
   "devDependencies": {
-    "@types/jest": "^29.5.11",
     "buffer-lite": "^2.1.0",
-    "jest": "^29.7.0",
-    "proxy-deep": "^3.1.1",
-    "ts-jest": "^29.1.2",
-    "tslib": "^2.3.1",
-    "typescript": "^5.3.3"
+    "proxy-deep": "^3.1.1"
   },
   "publishConfig": {
     "access": "public"
